quickdict/requirements.txt
PyQt6>=6.5
pynput>=1.7
pyautogui>=0.9
uiautomation>=2.0
rapidocr-onnxruntime>=1.3
opencv-python>=4.8
pyttsx3>=2.90
Pillow>=10.0
